Move WP.com API client creation from PostRsRestClient to WpApiClientProvider
Centralizes WP.com WpApiClient caching and creation in WpApiClientProvider
(fluxc module), eliminating duplicate dependencies and sign-out cleanup
from PostRsRestClient. The shared createWpComAuthProvider helper also
moves to fluxc so both WpApiClientProvider and WpServiceProvider can use it.
